+++ src/inspect_ai/_view/www/src/constants.ts
@@ -45,12 +45,3 @@ export const kScoreTypeOther = "other";
 export const kScoreTypeObject = "object";
 export const kScoreTypeBoolean = "boolean";
 export const kScoreTypeList = "list";
-
-// Sorting constants
-export const kSampleAscVal = "sample-asc";
-export const kSampleDescVal = "sample-desc";
-export const kEpochAscVal = "epoch-asc";
-export const kEpochDescVal = "epoch-desc";
-export const kScoreAscVal = "score-asc";
-export const kScoreDescVal = "score-desc";
-export const kDefaultSort = kSampleAscVal;